==> Notebooks/03 TCC - Tratamento Data Frame Faltas.ipynb <==
# 3.2.2 - Pré Processamento de Dados: Dataframe Faltas

#Importando Pandas, OS para leitura do arquivo e Matplotlib

import pandas as pd
import os
import matplotlib.pyplot as plt

os.chdir('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro')

#Importando o data set dos dados das partidas dos campeonatos brasileiros e os dados dos cartões amarelos

df_jogos=pd.read_csv('campeonato-brasileiro-full.csv',sep=',')
df_faltas=pd.read_csv('campeonato-brasileiro-estatisticas-full.csv',sep=',')


#Foi identificada uma falta de dados no Data Frame de estatisticas, onde os dados de cartões e faltas são registrados após
#o registro da partida 4987. Já no Data Frame de cartões é identificado que os registros são contabilizados a partir da 
# partida com o ID 4607. Essa diferença é basicamente de um campeonato inteiro sem dados que poderiam ser aproveitado
#por esse motivo iremos tratar o Data Frame de cartões para utilizar os dados a partir deste Data Frame.

#Inicio Campeonato 4985 - 6885

df_faltas=df_faltas[(df_faltas.partida_id>=4987) & (df_faltas.partida_id<=6885)]
df_faltas[(df_faltas.partida_id>=5500)]

#Identificando valores nulos ou zerados no Data Frame de faltas
#Primeiramente verificamos o tamanho do Data Frame
df_jogos.info()


#Idenficando valores nulos
df_faltas.info()

#Visualizando o Data Frame
df_faltas.head(100)

#Para identificar os registros que possuem os valores de cartões zerados de fato nas partidas será necessário um tratamento no
#Dataframe fazendo a transposição das linhas das partidas para uma linha única.
df_faltas_temp = df_faltas.copy(deep=True)
df_faltas_temp.head()

#Criando a coluna Mandante no DataFrame de faltas temporário e atribuindo um valor aleatório
df_faltas_temp['mandante'] = 3
#df_faltas_temp

#Percorrendo o Data Frame de Faltas
for linha in range(len(df_faltas_temp)):

    #Coletando o Id 
    id=df_faltas_temp.iloc[linha,0].astype(str)

    #Buscando o mandante do confronto pelo ID no Data Frame dos jogos
    mandante_jogos = df_jogos[['ID','mandante']].query('ID=='+id)
    mandante_jogos = mandante_jogos.iloc[0,1]

    #Coletando o time no Data Frame dos faltas
    mandante_faltas = df_faltas_temp.iloc[linha,2]
    
    #Identificando se o clube é mandante, caso seja o mesmo clube significa que o clube é mandante caso contrário o clube é
    #visitante
    
    if mandante_faltas==mandante_jogos:
        #Atribui o valor 1 na Coluna Mandante
        df_faltas_temp['mandante'].iloc[linha] = 1
        
    else:
        df_faltas_temp['mandante'].iloc[linha] = 0


#Conferindo se a coluna de mandante foi atribuida corretamente
df_jogos[df_jogos.ID==8011]

#Criando um DataFrame com os registros do DataFrame de cartões amarelos em um DataFrame com as linhas transpostas
df_faltas_tratados_temp = df_faltas_temp
df_faltas_tratados_temp= df_faltas_tratados_temp[['partida_id','clube','cartao_amarelo','cartao_vermelho','passes','faltas','mandante']]

mask = df_faltas_tratados_temp['mandante'] == 1

df_faltas_tratados_temp = df_faltas_tratados_temp.drop(columns = ['mandante'])

mandante = df_faltas_tratados_temp.loc[mask]
visitante = df_faltas_tratados_temp.loc[~mask]

colunas_mandante = ['partida_id','clube','cartao_amarelo','cartao_vermelho','passes','faltas']
colunas_visitante = ['partida_id','clube','cartao_amarelo','cartao_vermelho','passes','faltas']


mandante.columns = colunas_mandante
visitante.columns = colunas_visitante

df_faltas_tratados = mandante.merge(visitante, on='partida_id')
df_faltas_tratados.head()

#Renomeando as colunas

df_faltas_tratados.rename(columns={'clube_x':'clube_mandante'},inplace=True)
df_faltas_tratados.rename(columns={'clube_y':'clube_visitante'},inplace=True)
df_faltas_tratados.rename(columns={'faltas_x':'faltas_mandante'},inplace=True)
df_faltas_tratados.rename(columns={'faltas_y':'faltas_visitante'},inplace=True)
df_faltas_tratados.rename(columns={'cartao_amarelo_x':'cartao_amarelo_mandante'},inplace=True)
df_faltas_tratados.rename(columns={'cartao_amarelo_y':'cartao_amarelo_visitante'},inplace=True)
df_faltas_tratados.rename(columns={'cartao_vermelho_x':'cartao_vermelho_mandante'},inplace=True)
df_faltas_tratados.rename(columns={'cartao_vermelho_y':'cartao_vermelho_visitante'},inplace=True)
df_faltas_tratados.rename(columns={'passes_x':'passes_mandante'},inplace=True)
df_faltas_tratados.rename(columns={'passes_y':'passes_visitante'},inplace=True)

#Criando a coluna de total de faltas da partida
df_faltas_tratados['total_faltas_partida']=df_faltas_tratados[['faltas_mandante','faltas_visitante']].sum(axis=1)

#Criando a coluna de total de passes da partida
df_faltas_tratados['total_passes_partida']=df_faltas_tratados[['passes_mandante','passes_visitante']].sum(axis=1)

#Criando a coluna de total de cartoes amarelos
df_faltas_tratados['total_cartoes_amarelos']=df_faltas_tratados[['cartao_amarelo_mandante','cartao_amarelo_visitante']].sum(axis=1)

#Criando a coluna de total de cartoes vermelhos
df_faltas_tratados['total_cartoes_vermelhos']=df_faltas_tratados[['cartao_vermelho_mandante','cartao_vermelho_visitante']].sum(axis=1)

#Criando a coluna de total de cartoes da partida
df_faltas_tratados['total_cartoes_partida']=df_faltas_tratados[['total_cartoes_amarelos','total_cartoes_vermelhos']].sum(axis=1)


df_faltas_tratados.tail()

#Identificando as partidas que não possuem cartões, porém, que não são dados faltantes
df_faltas_tratados[(df_faltas_tratados.total_cartoes_partida==0) & (df_faltas_tratados.total_faltas_partida!=0) & (df_faltas_tratados.total_passes_partida!=0)]
df_faltas_tratados[df_faltas_tratados.partida_id==6828]


#Conferindo o DataFrame
df_faltas_tratados_temp.head()

#Salvando o Dataframe em um diretório local
df_faltas_tratados = df_faltas_tratados_temp
df_faltas_tratados.to_csv('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro\\df_faltas_tratados.csv')
