==> Notebooks/08 TCC - Seleção de Features.ipynb <==
# 5.1 - Features Utilizadas para os Modelos

import pandas as pd
import os
import matplotlib.pyplot as plt
from sklearn.feature_selection import SelectKBest

os.chdir('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro')

#Carregando os dados do Data Frame

df_dados=pd.read_csv('df_dados_fe.csv',sep=',')
df_dados.head()

#Verificando os melhores atributos para tentar prever o total de cartões na partida

features = df_dados.drop(['ID','campeonato','rodata','data','clube_mandante','clube_visitante','qtd_cartao_amarelo_mandante',
'qtd_cartao_amarelo_visitante','qtd_cartao_vermelho_mandante','qtd_cartao_vermelho_visitante',
'total_cartoes_vermelhos','total_cartoes_amarelos','total_cartoes_partida','chave_classico','total_cartoes_visitante',
'faltas_visitante','faltas_mandante','total_faltas_partida','Target_mandante_amarelo','data_dt'],axis=1)

features=features.drop(features.columns[0],axis=1)
label = df_dados[['Target_mandante_amarelo']]

#Verificando as melhores features com K_best

lista_features=features.columns.tolist()
k_best_features = SelectKBest(k='all')
k_best_features.fit_transform(features,label)
k_best_features_scores = k_best_features.scores_
raw_pairs = zip(lista_features[1:],k_best_features_scores)

ordered_pairs = list(reversed(sorted(raw_pairs, key=lambda x: x[1])))

k_best_features_final = dict(ordered_pairs[:15])
best_features = k_best_features_final.keys()
print('')
print('Melhores Features')
print(k_best_features_final)
