==> Notebooks/09 TCC - Normalização dos Dados.ipynb <==
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.model_selection import GridSearchCV
from sklearn import metrics
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.metrics import mean_squared_error, accuracy_score, f1_score,precision_score,recall_score
import seaborn as sn
import matplotlib.pyplot as plt
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC
from sklearn.feature_selection import SelectKBest
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from IPython.display import display
from sklearn.preprocessing import scale


os.chdir('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro')

#Carregando os dados do Data Frame

df_dados=pd.read_csv('df_dados_fe.csv',sep=',')
features = pd.read_csv('features.csv',sep=',')
labels = pd.read_csv('label.csv',sep=',')
df_dados.head()

#Verificar a dimensão do Dataframe

labels.shape

#Verificando os registros e colunas

df_dados.info()

df_dados_norm = df_dados.copy(deep=True)

#Código para normalizar os dados

from sklearn.preprocessing import MinMaxScaler

def normalize_group(group):
    return scaler.fit_transform(group.values.reshape(-1, 1)).flatten()


scaler = MinMaxScaler()

df_dados_norm['qtd_cartao_amarelo_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['qtd_cartao_amarelo_mandante'].transform(normalize_group)
df_dados_norm['qtd_cartao_amarelo_visitante'] = df_dados.groupby('clube_visitante')['qtd_cartao_amarelo_visitante'].transform(normalize_group)

df_dados_norm['qtd_cartao_vermelho_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['qtd_cartao_vermelho_mandante'].transform(normalize_group)
df_dados_norm['qtd_cartao_vermelho_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['qtd_cartao_vermelho_visitante'].transform(normalize_group)

df_dados_norm['faltas_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['faltas_mandante'].transform(normalize_group)
df_dados_norm['faltas_visitante'] = df_dados.groupby(['campeonato','clube_mandante'])['faltas_visitante'].transform(normalize_group)

df_dados_norm['MedAma5_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MedAma5_mandante'].transform(normalize_group)
df_dados_norm['MedVer5_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MedVer5_mandante'].transform(normalize_group)
df_dados_norm['MedAma5_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MedAma5_visitante'].transform(normalize_group)
df_dados_norm['MedVer5_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MedVer5_visitante'].transform(normalize_group)
df_dados_norm['MedAmaTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MedAmaTemp_mandante'].transform(normalize_group)
df_dados_norm['MedAmaTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MedAmaTemp_visitante'].transform(normalize_group)
df_dados_norm['MedVerTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MedVerTemp_mandante'].transform(normalize_group)
df_dados_norm['MedVerTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MedVerTemp_visitante'].transform(normalize_group)
df_dados_norm['MedAmaH_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MedAmaH_mandante'].transform(normalize_group)
df_dados_norm['MedAmaH_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MedAmaH_visitante'].transform(normalize_group)
df_dados_norm['MedVerH_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MedVerH_mandante'].transform(normalize_group)
df_dados_norm['MedVerH_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MedVerH_visitante'].transform(normalize_group)

df_dados_norm['MaxAmaTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MaxAmaTemp_mandante'].transform(normalize_group)
df_dados_norm['MinAmaTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['MinAmaTemp_mandante'].transform(normalize_group)
df_dados_norm['MaxAmaTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MaxAmaTemp_visitante'].transform(normalize_group)
df_dados_norm['MinAmaTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['MinAmaTemp_visitante'].transform(normalize_group)

df_dados_norm['ScoreMed_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['ScoreMed_mandante'].transform(normalize_group)
df_dados_norm['ScoreLevel_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['ScoreLevel_mandante'].transform(normalize_group)
df_dados_norm['ScoreMed_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['ScoreMed_visitante'].transform(normalize_group)
df_dados_norm['ScoreLevel_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['ScoreLevel_visitante'].transform(normalize_group)
df_dados_norm['ScoreMedTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['ScoreMedTemp_mandante'].transform(normalize_group)
df_dados_norm['ScoreLevelTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['ScoreLevelTemp_mandante'].transform(normalize_group)
df_dados_norm['ScoreMedTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['ScoreMedTemp_visitante'].transform(normalize_group)
df_dados_norm['ScoreLevelTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['ScoreLevelTemp_visitante'].transform(normalize_group)

df_dados_norm['FalAma_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['FalAma_mandante'].transform(normalize_group)
df_dados_norm['FalVer_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['FalVer_mandante'].transform(normalize_group)
df_dados_norm['FalAma_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['FalAma_visitante'].transform(normalize_group)
df_dados_norm['FalVer_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['FalVer_visitante'].transform(normalize_group)
df_dados_norm['FalAmaTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['FalAmaTemp_mandante'].transform(normalize_group)
df_dados_norm['FalVerTemp_mandante'] = df_dados.groupby(['campeonato','clube_mandante'])['FalVerTemp_mandante'].transform(normalize_group)
df_dados_norm['FalAmaTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['FalAmaTemp_visitante'].transform(normalize_group)
df_dados_norm['FalVerTemp_visitante'] = df_dados.groupby(['campeonato','clube_visitante'])['FalVerTemp_visitante'].transform(normalize_group)
 


#Salvando os dados do Dataframe em um diretório local

df_dados_norm.to_csv('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro\\dados_normalizados.csv')
