==> Notebooks/10 TCC - Regressão Logistica.ipynb <==
# 5.6 - Regressão Logística

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import LinearRegression, Ridge, Lasso
from sklearn.model_selection import GridSearchCV
from sklearn import metrics
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.metrics import mean_squared_error, accuracy_score, f1_score,precision_score,recall_score
import seaborn as sn
import matplotlib.pyplot as plt
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC
from sklearn.feature_selection import SelectKBest
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from IPython.display import display
from sklearn.preprocessing import scale


os.chdir('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro')

#Carregando os dados do Data Frame

df_dados=pd.read_csv('dados_normalizados.csv',sep=',')

df_dados.head()

#Criando o data frame para coletar os dados de avaliação dos modelos
colunas = ["tipo", "modelo", "acuracia", "precisao_teste", "precisao_geral", "recall_teste", "recall_geral"]
df_resultados = pd.DataFrame(columns=colunas)

df_dados_norm = df_dados.copy(deep=True)

#Selecionando as colunas que serão utilizadas no modelo
df_dados_norm_visitante=df_dados_norm[['classico_regional','MedAma5_mandante','MedAma5_visitante',
'MedAmaTemp_mandante','MedAmaTemp_visitante','MedAmaH_mandante',
'MedAmaH_visitante','MaxAmaTemp_mandante','MinAmaTemp_mandante',
'MaxAmaTemp_visitante','MinAmaTemp_visitante','ScoreMed_mandante','ScoreLevel_mandante','ScoreMedTemp_mandante',
                             'ScoreLevelTemp_mandante','FalAma_mandante','FalAmaTemp_mandante','FalAma_visitante',
                             'FalAmaTemp_visitante','ScoreMed_visitante','ScoreLevel_visitante','ScoreLevelTemp_mandante',
                            'ScoreMedTemp_visitante']]

target= df_dados_norm[['Target_visitante_amarelo']]
y=target.copy()
X=df_dados_norm_visitante.copy()

#Dividindo os Dataframes que serão utilzados no modelo

X_train=df_dados_norm_visitante[:2128]
X_test=df_dados_norm_visitante[2128:3040]
y_train=target[:2128]
y_test=target[2128:3040]

#Verificando as dimensões dos Dataframes

print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

#Código de para aplicar o Grid Search

param_grid = {'C':[0.001, 0.01, 0.1, 1, 10, 100, 1000]}

search = GridSearchCV(LogisticRegression(max_iter=1000), param_grid)

search.fit(X_train, y_train)
clf = search.best_estimator_
pred = clf.predict(X_test)
lg_acc = accuracy_score(y_test,pred)
f1 = f1_score(y_test,pred,average='macro')

print('Acurácia Logistic Regression:{}'.format(lg_acc))
print('F1 Score :{}'.format(f1))

print(clf)

#Código para treinamento do modelo e exibição dos resultados

logistic_regression = LogisticRegression(C=0.01, max_iter=1000)
logistic_regression.fit(X_train,y_train)
pred = logistic_regression.predict(X_test)

lg_acc = accuracy_score(y_test,pred)


print('Acurácia Logistic Regression:{}'.format(lg_acc))


print ('\n Matriz de Confusão')
print(confusion_matrix(y_test,logistic_regression.predict(X_test)))

print('\n Relatório Completo')
print(classification_report(y,logistic_regression.predict(X)))
print('\n Relatório Teste')
print(classification_report(y_test,clf.predict(X_test)))

#Inserção dos resultados no Dataframe para análise.

dados = [
    {"tipo": "Visitantes", "modelo": "Regressão Logística", "acuracia": lg_acc.round(4), "precisao_teste": precision_score(y_test,clf.predict(X_test),average='macro').round(4), "precisao_geral": precision_score(y,logistic_regression.predict(X),average='macro').round(4), "recall_teste": recall_score(y_test,clf.predict(X_test),average='macro').round(4), "recall_geral": recall_score(y,logistic_regression.predict(X),average='macro').round(4)},
]
df_resultados_visitantes = df_resultados.append(dados, ignore_index=True)
df_resultados_visitantes

df_dados_norm = df_dados.copy(deep=True)

#Selecionando as colunas que serão utilizadas no modelo

df_dados_norm_mandante=df_dados_norm[['classico_regional','MedAma5_mandante','MedAma5_visitante',
'MedAmaTemp_mandante','MedAmaTemp_visitante','MedAmaH_mandante',
'MedAmaH_visitante','MaxAmaTemp_mandante','MinAmaTemp_mandante',
'MaxAmaTemp_visitante','MinAmaTemp_visitante','ScoreMed_mandante','ScoreLevel_mandante','ScoreMedTemp_mandante',
                             'ScoreLevelTemp_mandante','FalAma_mandante','FalAmaTemp_mandante','FalAma_visitante',
                             'FalAmaTemp_visitante','ScoreMed_visitante','ScoreLevel_visitante','ScoreLevelTemp_mandante',
                            'ScoreMedTemp_visitante']]

target= df_dados_norm[['Target_mandante_amarelo']]
y=target.copy()
X=df_dados_norm_mandante.copy()

#Dividindo os Dataframes que serão utilzados no modelo

X_train=df_dados_norm_mandante[:2128]
X_test=df_dados_norm_mandante[2128:3040]
y_train=target[:2128]
y_test=target[2128:3040]

#Verificando as dimensões dos Dataframes

print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

#Código de para aplicar o Grid Search

param_grid = {'C':[0.001, 0.01, 0.1, 1, 10, 100, 1000]}

search = GridSearchCV(LogisticRegression(max_iter=1000), param_grid)

search.fit(X_train, y_train)
clf = search.best_estimator_
pred = clf.predict(X_test)
lg_acc = accuracy_score(y_test,pred)
f1 = f1_score(y_test,pred,average='macro')

print('Acurácia Logistic Regression:{}'.format(lg_acc))
print('F1 Score :{}'.format(f1))

print(clf)

#Código para treinamento do modelo e exibição dos resultados

logistic_regression = LogisticRegression(C=0.1, max_iter=1000)
logistic_regression.fit(X_train,y_train)
pred = logistic_regression.predict(X_test)

lg_acc = accuracy_score(y_test,pred)

print('Acurácia Logistic Regression:{}'.format(lg_acc))

print('\n Relatório Completo')
print(classification_report(y,logistic_regression.predict(X)))
print('\n Relatório Teste')
print(classification_report(y_test,clf.predict(X_test)))

#Inserção dos resultados no Dataframe para análise.

dados = [
    {"tipo": "Mandantes", "modelo": "Regressão Logística", "acuracia": lg_acc.round(4), "precisao_teste": precision_score(y_test,clf.predict(X_test),average='macro').round(4), "precisao_geral": precision_score(y,logistic_regression.predict(X),average='macro').round(4), "recall_teste": recall_score(y_test,clf.predict(X_test),average='macro').round(4), "recall_geral": recall_score(y,logistic_regression.predict(X),average='macro').round(4)},
]
df_resultados_mandantes = df_resultados.append(dados, ignore_index=True)
df_resultados_mandantes

#Salvando os Dataframes em diretório local

df_resultados_mandantes.to_csv('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro\\df_resultados_mandantes.csv')
df_resultados_visitantes.to_csv('D:\\Users\\Diogo\\Documents\\Pós Graduação\\Módulos\\TCC\\DataSets\\Futebol Brasileiro\\df_resultados_visitantes.csv')
